requirements.txt
google-adk[database]==1.7.0
google-generativeai==0.8.5
python-dotenv==1.1.0
deprecated